Extracts config_renderer file
